backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
nbformat>=5.10.4
nbconvert>=7.16.4
jupyter_client>=8.6.0
ipykernel>=6.29.5
codecarbon>=2.4.2
chromadb>=0.5.3
sentence-transformers>=3.0.1
groq>=0.9.0
reportlab>=4.2.0
matplotlib>=3.9.0
numpy>=1.26.4
psutil>=5.9.8
pydantic>=2.7.4
python-dotenv>=1.0.1
aiofiles>=23.2.1
jinja2>=3.1.4
pandas>=2.2.2
